backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-cryptography==1.1
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.7
psycopg2-binary==2.9.9

